requirements.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==11.0.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
django-crispy-forms>=2.3
djangorestframework==3.15.0
drf-yasg==1.21.7
crispy-bootstrap5==2025.6
